Move models folder inside lib
for clarity of references move models folder
