+++ lib/datamodel/entity.js
