+++ lib/datamodel/graph/edge.js
