+++ lib/datamodel/graph/index.js
