+++ lib/datamodel/graph/node.js
