+++ lib/datamodel/model-property.js
