+++ test/acceptance/use-cases/test-scripts/create-models.js
@@ -3,7 +3,7 @@ const app = require('../../../../');
 const expect = require('../../../helpers/expect');
 const fs = require('fs-extra');
 const loopback = require('loopback');
-const ModelClass = require('../../../../datamodel/model');
+const ModelClass = require('../../../../lib/datamodel/model');
 const path = require('path');
 const testSupport = require('../../../helpers/test-support');
 const util = require('util');
